// workshop-requisition-codeapp/src/generated/models/Wksp_suppliersModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Wksp_suppliersstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Wksp_suppliersstatecode = keyof typeof Wksp_suppliersstatecode;
export const Wksp_suppliersstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Wksp_suppliersstatuscode = keyof typeof Wksp_suppliersstatuscode;
export const Wksp_supplierswksp_contractstatus = {
  100000000: 'On-Contract',
  100000001: 'Off-Contract'
} as const;
export type Wksp_supplierswksp_contractstatus = keyof typeof Wksp_supplierswksp_contractstatus;

export interface Wksp_suppliersBase {
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Wksp_suppliersstatecode;
  statuscode?: Wksp_suppliersstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
  wksp_contactname?: string;
  wksp_contractstatus?: Wksp_supplierswksp_contractstatus;
  wksp_email?: string;
  wksp_phone?: string;
  wksp_supplierid: string;
  wksp_suppliername?: string;
}

export interface Wksp_suppliers extends Wksp_suppliersBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  wksp_contractstatusname?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}